optking and libsapt_solver, mostly print errors

+++ psi4/src/psi4/libsapt_solver/disp22t.cc
@@ -333,7 +333,7 @@ double SAPT2p::disp220t(int AAfile, const char *AAlabel, const char *ARlabel, co
         }
         stop = std::time(nullptr);
         if (print_) {
-            outfile->Printf("    (i = %3d of %3d) %10ld seconds\n", b + 1, aoccB, stop - start);
+            outfile->Printf("    (i = %3zu of %3zu) %10ld seconds\n", b + 1, aoccB, stop - start);
         }
     }
 
@@ -354,8 +354,8 @@ double SAPT2p::disp220t(int AAfile, const char *AAlabel, const char *ARlabel, co
 
 double SAPT2p::disp220tccd(int AAnum, const char *AA_label, int Rnum, const char *AR_label, const char *RR_label,
                            int BBnum, const char *BS_label, int ampnum, const char *tarar, const char *tbsar,
-                           double *evalsA, double *evalsB, int noccA, int nvirA, int foccA, int noccB, int nvirB,
-                           int foccB) {
+                           double *evalsA, double *evalsB, size_t noccA, size_t nvirA, size_t foccA, size_t noccB, size_t nvirB,
+                           size_t foccB) {
     double energy = 0.0;
 
     noccA -= foccA;
@@ -475,7 +475,7 @@ double SAPT2p::disp220tccd(int AAnum, const char *AA_label, int Rnum, const char
             }
         }
         stop = std::time(nullptr);
-        outfile->Printf("    (i = %3d of %3d) %10ld seconds\n", b + 1, noccB, stop - start);
+        outfile->Printf("    (i = %3zu of %3zu) %10ld seconds\n", b + 1, noccB, stop - start);
     }
 
     free(B_p_bs);
